Transfer code base from App.js to MadeOnEarth.js
