+++ src/App.css
@@ -9,16 +9,3 @@
 } */
 
 
-/* ===== NOTE TO SELF ===== */
-/* Display background image for screen width larger than mobile */
-@media screen and (min-width: 425px ) {
-  .app {
-    height: 100%;
-    width: 100%;
-    background-image: url('./asset/BackgroundIMG.jpg');
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-attachment: fixed;
-  }
-}
